requirements.txt
Flask==2.0.1
psycopg2==2.9.9
python-dotenv==0.19.0
gunicorn==20.1.0
streamlit==1.29.0
pandas==2.1.4 